fixed adding tags to db fixed entity objects fixed book_list_item.xml

+++ app/src/main/java/uk/co/droidinactu/elibrary/room/EBook.kt
@@ -24,10 +24,10 @@ class EBook() {
     var tags: MutableList<Tag> = ArrayList()
 
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0
+    var id: Long = 0
 
     @ColumnInfo
-    var inLibraryId = -1
+    var inLibraryId = -1L
 
     @ColumnInfo
     @NotNull
